backend/requirements/base.txt
daphne
django==4.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers
psycopg2-binary
channels==4.0.0
channels-redis==4.0.0
pillow
django-filter
drf_spectacular
django-environ
celery
gunicorn
twisted[http2,tls]
